+++ local_train.py
@@ -1,7 +1,7 @@
 import wandb
 
 from src.dataset_split import train, val, test, test_dummy, train_dummy, val_dummy
-from src.main import hyperparamter_tuning, training_testing, training_testing2
+from src.main import hyperparamter_tuning, training_testing
 
 
 if __name__ == '__main__':
@@ -21,6 +21,6 @@ if __name__ == '__main__':
         "weight_decay":1e-4, "trajectory_length":5, "dropout_p":0.5,
         "resize":64, "K":100, "epochs":10, "patience":40, "camera-correction":True}
 
-    training_testing2(args, wandb_project="deepvo-lightning-trials", visualization=True, wandb_name="trial_runs")
+    training_testing(args, wandb_project="deepvo-lightning-trials", visualization=True, wandb_name="trial_runs")
