+++ src/main.py
@@ -5,70 +5,14 @@ import os
 import wandb
 import time
 import datetime
+import matplotlib.pyplot as plt
 import sys
 
-from .dataset import DuckietownDataset
-from .utils import human_format, count_parameters
-from .model import DeepVONet
-from .training import plot_test, plot_train_valid, train_model, test_model
 import pytorch_lightning as pl
-
+from .training import DeepVONet
+from .utils import plot_test
 
 def training_testing(args, wandb_project, visualization=True, wandb_name=None):
-    # tell wandb to get started
-    run = wandb.init(project=wandb_project, entity="av_deepvo", name=wandb_name, config=args)
-    # access all HPs through wandb.config, so logging matches execution!
-    # wandb.config = args
-    print(args)
-    train_data = DuckietownDataset(args["train_split"], args)
-    val_data = DuckietownDataset(args["val_split"], args)
-    test_data = DuckietownDataset(args["test_split"], args)
-
-    train_loader = torch.utils.data.DataLoader(train_data, batch_size=args["bsize"], shuffle=True, drop_last=True)
-    val_loader = torch.utils.data.DataLoader(val_data, batch_size=args["bsize"], shuffle=True, drop_last=True)
-    test_loader = torch.utils.data.DataLoader(test_data, batch_size=1, shuffle=False, drop_last=True)
-
-    if not os.path.exists(args["checkpoint_path"]): os.makedirs(args["checkpoint_path"])
-
-    if args["model"] == "DeepVO":
-        model = ConvLstm(args["resize"], args["resize"], args["dropout_p"])
-    elif args["model"] == "DeepVOConv":
-        model = Conv(args["resize"], args["resize"], args["dropout_p"])
-    else:
-        print("model not defined")
-        sys.exit()
-
-    print("Number of parameters:", human_format(count_parameters(model)))
-    print("Number of parameter bytes:", human_format(32 * count_parameters(model)))
-
-    USE_GPU = torch.cuda.is_available()
-
-    if USE_GPU:
-        model.cuda()
-        print("Running model with cuda")
-
-    if args["checkpoint"] is not None:
-        print("loading old model")
-        the_checkpoint = torch.load(args["checkpoint"])
-        model.load_state_dict(the_checkpoint['state_dict'])
-
-    best_model, logs, args_ = train_model(model, train_loader, val_loader, args)
-    if visualization:
-        plot_train_valid(logs, args_)
-
-    print(logs)
-
-    test_loss, relative_poses_pred = test_model(best_model, test_loader, args)
-    if visualization:
-        plot_test(test_data, relative_poses_pred)
-
-    save_model_onnx(best_model, args)
-
-    run.finish()
-
-    return logs, test_loss.detach().numpy()
-
-def training_testing2(args, wandb_project, visualization=True, wandb_name=None):
     # experiment tracker (you need to sign in with your account)
 
     wandb.require(experiment="service")
@@ -100,7 +44,8 @@ def training_testing2(args, wandb_project, visualization=True, wandb_name=None):
             logger=wandb_logger,
             callbacks=checkpoint,
             max_epochs=args["epochs"],
-            log_every_n_steps=10
+            log_every_n_steps=10,
+            progress_bar_refresh_rate=20
         )
 
     model = DeepVONet(args)
@@ -112,6 +57,11 @@ def training_testing2(args, wandb_project, visualization=True, wandb_name=None):
 
     print("------- Testing Begins! -------")
     trainer.test(model)
+    plot_test(model.test_data, model.trajectories)
+    save_model_onnx(model, args)
+
+    wandb_logger.unwatch(model)
+
 
 
 def hyperparamter_tuning(args, wandb_project, visualization=False, wandb_name=None):
@@ -155,3 +105,5 @@ def save_model_onnx(model, args):
                       dynamic_axes={'input': {0: 'batch_size'},  # variable length axes
                                     'output': {0: 'batch_size'}})
     wandb.save(filename)
+
+
